@@ -17,15 +17,15 @@
     "@types/node": "^16.7.13",
     "@types/react": "^17.0.20",
     "@types/react-datepicker": "^4.3.4",
-    "@types/react-dom": "^17.0.9",
+    "@types/react-dom": "^18.0.3",
     "@types/styled-components": "^5.1.19",
     "axios": "^0.24.0",
     "date-fns": "^2.28.0",
     "dayjs": "^1.11.0",
-    "react": "^17.0.2",
+    "react": "^18.1.0",
     "react-datepicker": "^4.6.0",
     "react-daum-postcode": "^3.0.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.1.0",
     "react-kakao-maps-sdk": "^1.0.4",
     "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
