@@ -1,7 +1,7 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
-import { BrowserRouter } from 'react-router-dom'
+import ReactDOM from 'react-dom/client'
 import App from './App'
+import { BrowserRouter } from 'react-router-dom'
 import GlobalStyles from './assets/css/GlobalStyles'
 import axios from 'axios'
 import './assets/css/App.css'
@@ -11,7 +11,9 @@ axios.defaults.withCredentials = true
 axios.defaults.baseURL =
   process.env.NODE_ENV === 'production' ? process.env.REACT_APP_API_URL : process.env.REACT_APP_API_URL
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root') as HTMLDivElement)
+
+root.render(
   <React.StrictMode>
     <RecoilRoot>
       <GlobalStyles />
@@ -20,5 +22,4 @@ ReactDOM.render(
       </BrowserRouter>
     </RecoilRoot>
   </React.StrictMode>,
-  document.getElementById('root'),
 )
